// tests/fixtures/types/language-types.generated.ts
export type AceMode = "asciidoc" | "c_cpp" | "markdown" | "text";
export type Aliases = readonly "ags"[];
export type CodemirrorMimeType = "application/pgp" | "text/x-c++src";
export type CodemirrorMode = "asciiarmor" | "clike";
export type Color = "#2ACCA8" | "#34EB6B" | "#38761D" | "#73a0c5" | "#814CCC" | "#B9D9B9" | "#B9D9FF";
export type Extensions = readonly (
	| ".2da"
	| ".adoc"
	| ".aidl"
	| ".apib"
	| ".asc"
	| ".asciidoc"
	| ".ash"
	| ".bsl"
	| ".os"
	| ".pub"
)[];
export type Interpreters = readonly "aidl"[];
export type LanguageId = 0 | 2 | 5 | 22 | 298 | 387204628 | 451700185 | 852099832;
export type LanguageName =
	| "1C Enterprise"
	| "2-Dimensional Array"
	| "AGS Script"
	| "AIDL"
	| "API Blueprint"
	| "AsciiDoc"
	| "KoLmafia ASH"
	| "Public Key";
export type TmScope =
	| "none"
	| "source.2da"
	| "source.aidl"
	| "source.ash"
	| "source.bsl"
	| "source.c++"
	| "text.html.asciidoc"
	| "text.html.markdown.source.gfm.apib";
export type Type = "data" | "markup" | "programming" | "prose";
export type Wrap = true;

export type Language = {
	ace_mode: AceMode;
	aliases?: Aliases;
	codemirror_mime_type?: CodemirrorMimeType;
	codemirror_mode?: CodemirrorMode;
	color?: Color;
	extensions: Extensions;
	interpreters?: Interpreters;
	language_id: LanguageId;
	name: LanguageName;
	tm_scope: TmScope;
	type: Type;
	wrap?: Wrap;
};
// ...
export type Dictionary<T> = Record<string & {}, T>;